+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### [Latest]
 
+- Minor change to how n_classes are calculated in confusion matrix [!306](https://github.com/umami-hep/puma/pull/306)
+
 ### [v0.4.4] (2025/02/17)
 
 - Adding show() function support [!304](https://github.com/umami-hep/puma/pull/304)